Euclan/requirements.txt
numpy==1.26.2
Pillow==10.1.0
pandas==2.1.4
pytz==2023.3.post1
tzdata==2023.3
opencv-python==4.8.1.78
scikit-image==0.22.0
imageio==2.33.1
torch==2.1.2
datasets==2.16.0
tqdm==4.66.1
click==8.1.7
scipy==1.11.4
scikit-learn==1.3.2
joblib==1.3.2
threadpoolctl==3.2.0
matplotlib==3.8.2
wandb==0.16.1
coloredlogs==15.0.1
urllib3==2.1.0
charset-normalizer==3.3.2
idna==3.6
PyYAML==6.0.1
typing_extensions==4.9.0
